third_party/WebKit/LayoutTests/fast/dom/Element/offset-parent-transform.html

<!DOCTYPE html>
<div id="transform" style="transform:translateX(1px);">
    <div id="test"></div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    test(function() {
       assert_equals(document.getElementById("test").offsetParent, document.getElementById("transform"));
    }, "inside transform");
</script>
